@@ -6,9 +6,9 @@
   "packageManager": "yarn@1.22.18",
   "dependencies": {
     "axios": "^0.26.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-query": "^3.35.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-query": "^3.38.0",
     "ulid": "^2.3.0",
     "web-vitals": "^2.1.4"
   },
@@ -24,17 +24,17 @@
     "@storybook/manager-webpack5": "^6.4.22",
     "@storybook/node-logger": "^6.4.22",
     "@storybook/preset-create-react-app": "^4.1.0",
-    "@storybook/react": "^6.4.21",
-    "@storybook/testing-library": "^0.0.9",
+    "@storybook/react": "^6.4.22",
+    "@storybook/testing-library": "^0.0.11",
     "@storybook/testing-react": "^1.2.4",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.1.1",
     "@testing-library/react-hooks": "^8.0.0",
     "@testing-library/user-event": "^14.1.1",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.25",
-    "@types/react": "^17.0.44",
-    "@types/react-dom": "^17.0.15",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
     "chromatic": "^6.5.4",
     "msw": "^0.39.2",
     "msw-storybook-addon": "^1.6.3",
