@@ -1,10 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { StrictMode } from "react";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
+import { createRoot } from "react-dom/client";
 
 const prepare = () => {
   if (process.env.NODE_ENV === "development") {
@@ -17,14 +17,15 @@ const prepare = () => {
 const queryClient = new QueryClient();
 
 prepare().then(() => {
-  ReactDOM.render(
-    <React.StrictMode>
+  const container = document.getElementById("root");
+  const root = createRoot(container!);
+  root.render(
+    <StrictMode>
       <QueryClientProvider client={queryClient}>
         <App />
         <ReactQueryDevtools initialIsOpen={false} />
       </QueryClientProvider>
-    </React.StrictMode>,
-    document.getElementById("root")
+    </StrictMode>
   );
 });
 
