@@ -1,6 +1,6 @@
-import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactNode } from "react";
-import { act, renderHook } from "@testing-library/react-hooks";
+import { QueryClient, QueryClientProvider, setLogger, UseQueryResult } from "react-query";
+import { act, renderHook, waitFor } from "@testing-library/react";
 import {
   useMutationCompleteAllTodos,
   useMutationDeleteCompletedTodos,
@@ -60,262 +60,214 @@ let testTodos: Todo[] = [
   },
 ];
 
-// QueryClientをテスト対象のカスタムHookをラップする
-const queryClientWrapper = () => {
-  const queryClient = new QueryClient();
-  return ({ children }: { children: ReactNode }) => (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  );
-};
+// QueryClientオブジェクトをテスト用の設定にして生成
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+      cacheTime: Infinity,
+      staleTime: Infinity,
+    },
+  },
+});
+
+// Loggingは ReactQuery v4からQueryClientのDefaultOptionsに組み込まれる
+setLogger({
+  log: console.log,
+  warn: console.warn,
+  error: process.env.NODE_ENV === "development" ? () => {} : console.error,
+});
+
+// テスト対象のカスタムHookでQueryClientを利用するためのカスタムWrapper
+const wrapper = ({ children }: { children: ReactNode }) => (
+  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+);
 
 describe("React QueryによるServerState管理", () => {
   beforeEach(() => {
-    // Given: MSWで利用するTodoデータを入れ替える
+    // テストごとにQueryClientのキャッシュをクリアする
+    queryClient.clear();
+    // Todoデータをリセットする
     setMockedTodo(testTodos);
   });
 
   describe("useTodoQueryのテスト", () => {
     test("バックエンドAPIを使いTodoリストが取得できること", async () => {
       // When: 全てのTodoを検索する
-      const { result, waitFor } = renderHook(() => useQueryTodo(), {
-        wrapper: queryClientWrapper(),
-      });
-      await waitFor(() => result.current.isSuccess);
+      const { result } = renderHook<UseQueryResult, unknown>(
+        () => useQueryTodo<Todo[]>(),
+        { wrapper }
+      );
+      await waitFor(() => expect(result.current.isSuccess).toBeTruthy());
 
       // Then: todoが取得できること
       expect(result.current.data).toHaveLength(7);
       expect(result.current.data).toEqual(testTodos);
     });
   });
+
   describe("useMutationTodoのテスト", () => {
     test("Todoを追加するとTodoリストを再フェッチする", async () => {
-      // Given: useQueryTodo と useMutationTodoAdded が参照するqueryClientを生成する
-      const wrapper = queryClientWrapper();
-
-      // カスタムHook useQueryTodoを出力しTodoリストをFetchする
-      const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-        () => useQueryTodo(),
-        {
-          wrapper: wrapper,
-        }
+      // Given: 検証用のuseQueryTodoカスタムフックを出力する
+      const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+        () => useQueryTodo<Todo[]>(),
+        { wrapper }
       );
-      await waitForQuery(() => resultQuery.current.isSuccess);
-      // 最初は７件
+      await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
       expect(resultQuery.current.data).toHaveLength(7);
 
-      // When: useMutation カスタムHookを出力する
-      const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+      // When: mutateを実行して新しいTodoを実行する
+      const { result: resultMutation } = renderHook(
         () => useMutationTodoAdded(),
-        {
-          wrapper: wrapper,
-        }
+        { wrapper }
+      );
+      const text = "new todo";
+      await act(() => resultMutation.current.mutate({ text }));
+      await waitFor(() =>
+        expect(resultMutation.current.isSuccess).toBeTruthy()
       );
 
-      // mutateを実行して新しいTodoを実行する
-      const text = "hogehoge";
-      act(() => {
-        resultMutation.current.mutate({ text });
-      });
-      await waitForMutation(() => resultMutation.current.isSuccess, {
-        timeout: 1500,
-      });
-
-      // Then: queryデータが再FetchされTodoが追加される
+      // Then: resultQueryが再FetchされTodoが追加される
       const todosAfterAdded: Todo[] = resultQuery.current.data as Todo[];
       expect(todosAfterAdded).toHaveLength(8);
       expect(todosAfterAdded[7].text).toBe(text);
       expect(todosAfterAdded[7].isCompleted).toBeFalsy();
       expect(todosAfterAdded[7].color).toBe(TODO_COLOR.None);
 
-      // responseに含まれる追加されたTodoオブジェクトを確認する
-      const addedTodo: Todo = resultMutation.current.data?.data;
-      expect(addedTodo.text).toBe(text);
-      expect(addedTodo.isCompleted).toBeFalsy();
-      expect(addedTodo.color).toBe(TODO_COLOR.None);
+      // 登録したtextが、TodoオブジェクトでresultMutationに返される
+      const after: Todo = resultMutation.current.data?.data;
+      expect(after.text).toBe(text);
+      expect(after.isCompleted).toBeFalsy();
+      expect(after.color).toBe(TODO_COLOR.None);
     });
 
     test("指定したIDのTodoをCompletedにする", async () => {
-      // Given: useQueryTodo と useMutationTodoAdded が参照するqueryClientを生成する
-      const wrapper = queryClientWrapper();
-
-      // カスタムHook useQueryTodoを出力しTodoリストをFetchする
-      const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-        () => useQueryTodo(),
-        {
-          wrapper: wrapper,
-        }
+      // Given: 検証用のuseQueryTodoカスタムフックを出力する
+      const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+        () => useQueryTodo<Todo[]>(),
+        { wrapper }
       );
-      await waitForQuery(() => resultQuery.current.isSuccess);
       // 最初は未完了であることを確認する
-      const todosBeforeMutation: Todo[] = resultQuery.current.data as Todo[];
-      expect(todosBeforeMutation[2].isCompleted).toBeFalsy();
+      await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
+      const before: Todo[] = resultQuery.current.data as Todo[];
+      expect(before[2].isCompleted).toBeFalsy();
 
-      // When: useMutation カスタムHookを出力する
-      const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+      // When: mutateを実行して指定IDのTodoのCompleted状況を変更する
+      const { result: resultMutation } = renderHook(
         () => useMutationTodoCompleted(),
-        {
-          wrapper: wrapper,
-        }
+        { wrapper }
       );
-
-      // mutateを実行して指定IDのTodoのCompleted状況を変更する
-      act(() => {
-        resultMutation.current.mutate({ id: "3" });
-      });
-      await waitForMutation(() => resultMutation.current.isSuccess, {
-        timeout: 1500,
-      });
+      await act(() => resultMutation.current.mutate({ id: "3" }));
+      await waitFor(() =>
+        expect(resultMutation.current.isSuccess).toBeTruthy()
+      );
 
       // Then: queryデータが再FetchされTodoの完了状況が更新される
-      const todosAfterMutation: Todo[] = resultQuery.current.data as Todo[];
-      expect(todosAfterMutation[2].text).toBe(todosBeforeMutation[2].text);
-      expect(todosAfterMutation[2].isCompleted).not.toBe(
-        todosBeforeMutation[2].isCompleted
-      );
-      expect(todosAfterMutation[2].color).toBe(todosBeforeMutation[2].color);
+      const after: Todo[] = resultQuery.current.data as Todo[];
+      expect(after[2].text).toBe(before[2].text);
+      expect(after[2].isCompleted).not.toBe(before[2].isCompleted);
+      expect(after[2].color).toBe(before[2].color);
     });
   });
 
   test("指定したIDのTodoのColorを変更する", async () => {
-    // Given: useQueryTodo と useMutationTodoAdded が参照するqueryClientを生成する
-    const wrapper = queryClientWrapper();
-
-    // カスタムHook useQueryTodoを出力しTodoリストをFetchする
-    const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-      () => useQueryTodo(),
-      {
-        wrapper: wrapper,
-      }
+    // Given: 検証用のuseQueryTodoカスタムフックを出力する
+    const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+      () => useQueryTodo<Todo[]>(),
+      { wrapper }
     );
-    await waitForQuery(() => resultQuery.current.isSuccess);
     // 最初はRedであることを確認する
-    const todosBeforeMutation: Todo[] = resultQuery.current.data as Todo[];
-    expect(todosBeforeMutation[1].color).toBe(TODO_COLOR.Red);
+    await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
+    const before: Todo[] = resultQuery.current.data as Todo[];
+    expect(before[1].color).toBe(TODO_COLOR.Red);
 
-    // When: useMutation カスタムHookを出力する
-    const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+    // When: mutateを実行してColorを変更する
+    const { result: resultMutation } = renderHook(
       () => useMutationTodoChangedColor(),
-      {
-        wrapper: wrapper,
-      }
+      { wrapper }
     );
-
-    // mutateを実行してColorを変更する
-    act(() => {
-      resultMutation.current.mutate({ id: "2", color: TODO_COLOR.Purple });
-    });
-    await waitForMutation(() => resultMutation.current.isSuccess, {
-      timeout: 1500,
-    });
+    await act(() =>
+      resultMutation.current.mutate({ id: "2", color: TODO_COLOR.Purple })
+    );
+    await waitFor(() => expect(resultMutation.current.isSuccess).toBeTruthy());
 
     // Then: queryデータが再FetchされTodoの完了状況が更新される
-    const todosAfterMutation: Todo[] = resultQuery.current.data as Todo[];
-    expect(todosAfterMutation[1].text).toBe(todosBeforeMutation[1].text);
-    expect(todosAfterMutation[1].isCompleted).toBe(
-      todosBeforeMutation[1].isCompleted
-    );
-    expect(todosAfterMutation[1].color).toBe(TODO_COLOR.Purple);
+    const after: Todo[] = resultQuery.current.data as Todo[];
+    expect(after[1].text).toBe(before[1].text);
+    expect(after[1].isCompleted).toBe(before[1].isCompleted);
+    expect(after[1].color).toBe(TODO_COLOR.Purple);
   });
 
   test("指定したIDのTodoをリストから削除する", async () => {
-    // Given: useQueryTodo と useMutationTodoAdded が参照するqueryClientを生成する
-    const wrapper = queryClientWrapper();
-
-    // カスタムHook useQueryTodoを出力しTodoリストをFetchする
-    const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-      () => useQueryTodo(),
-      {
-        wrapper: wrapper,
-      }
+    // Given: 検証用のuseQueryTodoカスタムフックを出力する
+    const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+      () => useQueryTodo<Todo[]>(),
+      { wrapper }
     );
-    await waitForQuery(() => resultQuery.current.isSuccess);
     // 最初は７件取得できる
+    await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
     expect(resultQuery.current.data).toHaveLength(7);
 
-    // When: useMutation カスタムHookを出力する
-    const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+    // When:mutateを実行して指定したIDを削除する
+    const { result: resultMutation } = renderHook(
       () => useMutationTodoDeleted(),
-      {
-        wrapper: wrapper,
-      }
+      { wrapper }
     );
-
-    // mutateを実行して指定したIDを削除する
-    act(() => {
-      resultMutation.current.mutate({ id: "4" });
-    });
-    await waitForMutation(() => resultMutation.current.isSuccess, {
-      timeout: 1500,
-    });
+    await act(() => resultMutation.current.mutate({ id: "4" }));
+    await waitFor(() => expect(resultMutation.current.isSuccess).toBeTruthy());
 
     // Then: queryデータが再FetchされTodoが６件になっていること
-    const todos: Todo[] = resultQuery.current.data as Todo[];
-    expect(todos).toHaveLength(6);
-    expect(todos[2].id).toBe("3");
-    expect(todos[3].id).toBe("5");
+    const after: Todo[] = resultQuery.current.data as Todo[];
+    expect(after).toHaveLength(6);
+    expect(after[2].id).toBe("3");
+    expect(after[3].id).toBe("5");
   });
 
   test("全てのTodoを完了済みにする", async () => {
-    // Given: Todoリストを作成する
-    const wrapper = queryClientWrapper();
-    const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-      () => useQueryTodo(),
-      {
-        wrapper: wrapper,
-      }
+    // Given: 検証用のuseQueryTodoカスタムフックを出力する
+    const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+      () => useQueryTodo<Todo[]>(),
+      { wrapper }
     );
-    await waitForQuery(() => resultQuery.current.isSuccess);
+    // 最初は７件取得できる
+    await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
     expect(resultQuery.current.data).toHaveLength(7);
 
-    // When: mutationを実行する
-    const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+    // When:mutateを実行して全てのTodoリストを完了済みにする
+    const { result: resultMutation } = renderHook(
       () => useMutationCompleteAllTodos(),
-      { wrapper: wrapper }
+      { wrapper }
     );
-
-    act(() => {
-      resultMutation.current.mutate();
-    });
-    await waitForMutation(() => resultMutation.current.isSuccess, {
-      timeout: 1500,
-    });
+    await act(() => resultMutation.current.mutate());
+    await waitFor(() => expect(resultMutation.current.isSuccess).toBeTruthy());
 
     // Then: TodoListの件数は最初と変わらず、全て完了になっていること
-    const todos: Todo[] = resultQuery.current.data as Todo[];
-    expect(todos).toHaveLength(7);
-    for (let todo of todos) {
-      expect(todo.isCompleted).toBeTruthy();
-    }
+    const after: Todo[] = resultQuery.current.data as Todo[];
+    expect(after).toHaveLength(7);
+    after.forEach((todo: Todo) => expect(todo.isCompleted).toBeTruthy());
   });
 
   test("完了済みのTodoをリストから削除する", async () => {
-    // Given: Todoリストを作る
-    const wrapper = queryClientWrapper();
-    const { result: resultQuery, waitFor: waitForQuery } = renderHook(
-      () => useQueryTodo(),
-      { wrapper: wrapper }
+    // Given: 検証用のuseQueryTodoカスタムフックを出力する
+    const { result: resultQuery } = renderHook<UseQueryResult, unknown>(
+      () => useQueryTodo<Todo[]>(),
+      { wrapper }
     );
-    await waitForQuery(() => resultQuery.current.isSuccess);
+    // 最初は７件取得できる
+    await waitFor(() => expect(resultQuery.current.isSuccess).toBeTruthy());
     expect(resultQuery.current.data).toHaveLength(7);
 
-    // When: useMutationを実行する
-    const { result: resultMutation, waitFor: waitForMutation } = renderHook(
+    // When: useMutationを実行して完了済みのTodoを削除する
+    const { result: resultMutation } = renderHook(
       () => useMutationDeleteCompletedTodos(),
-      {
-        wrapper: wrapper,
-      }
+      { wrapper }
     );
-
-    act(() => resultMutation.current.mutate());
-    await waitForMutation(() => resultMutation.current.isSuccess, {
-      timeout: 1500,
-    });
+    await act(() => resultMutation.current.mutate());
+    await waitFor(() => expect(resultMutation.current.isSuccess).toBeTruthy());
 
     // Then: TodoListには未完了のTodoだけが残る
-    const todos: Todo[] = resultQuery.current.data as Todo[];
-    expect(todos).toHaveLength(4);
-    for (let todo of todos) {
-      expect(todo.isCompleted).toBeFalsy();
-    }
+    const after: Todo[] = resultQuery.current.data as Todo[];
+    expect(after).toHaveLength(4);
+    after.forEach((todo: Todo) => expect(todo.isCompleted).toBeFalsy());
   });
 });
